+++ spring-webmvc/src/test/java/org/springframework/web/servlet/tags/form/InputTagTests.java
@@ -55,7 +55,7 @@ public class InputTagTests extends AbstractFormTagTests {
 		// set up test data
 		this.rob = new TestBean();
 		this.rob.setName("Rob");
-		this.rob.setMyFloat(Float.valueOf(12.34f));
+		this.rob.setMyFloat(12.34f);
 
 		TestBean sally = new TestBean();
 		sally.setName("Sally");
